+++ app/src/main/AndroidManifest.xml
@@ -44,7 +44,8 @@
             android:name=".httprequest"
             android:exported="false" />
         <service android:name=".GetListingsService" />
-        <service android:name = ".GetRatingsService"/>
+        <service android:name = ".RatingService$CreateRatingService"/>
+        <service android:name=".RatingService$GetRatingService"/>
         <service
             android:name=".RegService"/>
         <service
